moved classes related to the Transaction system into their own package
