+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/MarketHandler.java
@@ -11,6 +11,7 @@
 
 package com.internetEnemies.combatCritters.Logic;
 
+import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
 import com.internetEnemies.combatCritters.data.Database;
 import com.internetEnemies.combatCritters.data.IMarketDB;
 import com.internetEnemies.combatCritters.data.OffersDatabase;
