+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/transaction/TransactionVerify.java
@@ -1,4 +1,4 @@
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.transaction;
 
 /**
  * TransactionHandler.java
