+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/transaction/builders/MarketTransactionBuilder.java
@@ -1,4 +1,4 @@
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.transaction.builders;
 
 import com.internetEnemies.combatCritters.objects.Currency;
 import com.internetEnemies.combatCritters.objects.IMarketTransactionBuilder;
