+++ app/src/main/java/com/internetEnemies/combatCritters/data/MarketDB.java
@@ -11,18 +11,11 @@
 
 package com.internetEnemies.combatCritters.data;
 
-import com.internetEnemies.combatCritters.Logic.MarketTransactionBuilder;
-import com.internetEnemies.combatCritters.objects.Card;
-import com.internetEnemies.combatCritters.objects.ItemStack;
 import com.internetEnemies.combatCritters.objects.MarketTransaction;
-import com.internetEnemies.combatCritters.objects.Pack;
-import com.internetEnemies.combatCritters.objects.Transaction;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.TreeMap;
 
 public class MarketDB implements IMarketDB {
     private final List<MarketTransaction> bundleOffers;
