+++ app/src/main/java/com/internetEnemies/combatCritters/data/hsqldb/DSOHelpers/TransactionHelper.java
@@ -1,8 +1,8 @@
 package com.internetEnemies.combatCritters.data.hsqldb.DSOHelpers;
 
-import com.internetEnemies.combatCritters.Logic.MarketTransactionBuilder;
-import com.internetEnemies.combatCritters.Logic.TradeTransactionBuilder;
-import com.internetEnemies.combatCritters.Logic.RewardTransactionBuilder;
+import com.internetEnemies.combatCritters.Logic.transaction.builders.MarketTransactionBuilder;
+import com.internetEnemies.combatCritters.Logic.transaction.builders.TradeTransactionBuilder;
+import com.internetEnemies.combatCritters.Logic.transaction.builders.RewardTransactionBuilder;
 import com.internetEnemies.combatCritters.objects.Currency;
 import com.internetEnemies.combatCritters.objects.IItem;
 import com.internetEnemies.combatCritters.objects.ItemStack;
