+++ app/src/main/java/com/internetEnemies/combatCritters/data/hsqldb/TradeTransactionRegistryHSQLDB.java
@@ -1,6 +1,6 @@
 package com.internetEnemies.combatCritters.data.hsqldb;
 
-import com.internetEnemies.combatCritters.Logic.TradeTransactionBuilder;
+import com.internetEnemies.combatCritters.Logic.transaction.builders.TradeTransactionBuilder;
 import com.internetEnemies.combatCritters.data.IRegistry;
 import com.internetEnemies.combatCritters.data.hsqldb.DSOHelpers.TransactionHelper;
 import com.internetEnemies.combatCritters.objects.ItemStack;
