+++ app/src/test/java/com/internetEnemies/combatCritters/IntegrationTests/BattleRegistryIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.internetEnemies.combatCritters.IntegrationTests;
 
-import com.internetEnemies.combatCritters.Logic.ITransactionHandler;
-import com.internetEnemies.combatCritters.Logic.TransactionHandler;
+import com.internetEnemies.combatCritters.Logic.transaction.ITransactionHandler;
+import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
 import com.internetEnemies.combatCritters.Logic.battles.cards.BattleCardFactory;
 import com.internetEnemies.combatCritters.Logic.battles.events.EventSystem;
 import com.internetEnemies.combatCritters.Logic.battles.registry.BattleRegistry;
