+++ app/src/test/java/com/internetEnemies/combatCritters/LogicUnitTests/MarketCycleTest.java
@@ -5,7 +5,7 @@ import com.internetEnemies.combatCritters.Logic.IMarketCycle;
 import com.internetEnemies.combatCritters.Logic.IMarketHandler;
 import com.internetEnemies.combatCritters.Logic.MarketCycle;
 import com.internetEnemies.combatCritters.Logic.MarketHandler;
-import com.internetEnemies.combatCritters.Logic.MarketTransactionBuilder;
+import com.internetEnemies.combatCritters.Logic.transaction.builders.MarketTransactionBuilder;
 import com.internetEnemies.combatCritters.data.MarketDB;
 import com.internetEnemies.combatCritters.objects.Currency;
 import com.internetEnemies.combatCritters.objects.MarketTransaction;
